Make UID optional in frame submission
A non-existing UID means to run the job as the RQD user on the render
host; RQD does not attempt to switch users. (This has security
implications but probably not worse than the current setup which allows
any UID to be specified.)

This change allows Windows to submit jobs for Linux, as Cuesubmit will
now never send UIDs from Windows.

Changes:
- Alter database to allow UID to be `NULL` (not done to Oracle as
support is going to be dropped)
- Alter DAO classes to understand missing UID.
- Alter domain classes to allow UID to be `Optional<Integer>`.
- Update CJSL DTD to allow UID to be optional.
- Update .proto files to allow UID to be missing. According to Proto3
docs, this is binary-compatible. However, it is not backward-compatible
at the application level. Old RQDs will see a missing UID as UID 0
and refuse to run the frame.
- Update Cuesubmit to not send UIDs from Windows.