Simplify docker compose setup.
Place docker-compose at the root of the project.

Remove some of the variable user needed to setup with defaults.
As we can asume a person is running this as a test and not in production
with docker-compose.

process to start is now.

`docker-compose up` in the root of the project.