Keep staggerFrames interleaving start frame static
There was a implementation bug with the logic of the
start frame incrementing after each Nth step iteration,
causing the interleaving to be incorrect because the
step was operating on a subset of the frame range and
not the entire range for each Nth step. The start index
needs to remain static after each iter for interleaving
to work correctly on the entire range. Ex, 1-10:5. N=5,
results in 1,6,3,5,7,9,2,4,8,10 1st iter: N=5, [1,6],
2nd iter: N=2, [1,6,3,5,7,9], 3rd iter: N=1
[1,6,3,5,7,9,2,4,8,10].

(cherry picked from commit fd546f519f3d7d2d4985348c1bb3633dc93f757e)