Make sure nestedJobWhiteboard is processed in order
There was a problem where depending on the order of the hashmap
insertion, nested folders would not show up in the GUI. The issue
was caused by using childrenMap.entrySet which doesn't concern
about order.

To solve this issue the map values were encapsulated in a object
that implements the right sorting based on level, making sure
children paths are processed before their parents.