Replace BookingQueue by HealthyQueue
BookingQueue will not inherit from ThreadPoolExecutor, instead it will
manage an instance of HealthThreadPool, which is an extension of
threadPoolExecutor that handles healthChecks, termination and repeated
tasks. With this the Booking queue should be able to self-heal in case
of locked threads or database locks.

(cherry picked from commit 3b7ffd80930abdbd4896379e359ce5c001936b66)