Change dropDepends unittests for github
Cuegui MenuActions_tests.py's test_dropDepends
unittest no longer applies since the fn no longer
calls the pycue's dropDepends with a target
parameter. This is a bug fix to dropDepends and
now the fn iterates over each dependency and calls
"depend.satisfy" instead to remove since the
pycue'sdropDepends only applys to "one" depend
target, so it will not remove different multiple
different dependency types linked to the one frame.
Hence, this caused the Cuegui unittest to fail on
the Github PR. Opting to put the dropDepends test
in pycue (which didn't have one) and remove the test
in Cuegui, since Cuegui already has a test coverage
for depend.satisfy.