Update CueJobMonitorTree only with new data
Fixed processUpdate to only update in mem
dict self._items when "new" jobs are retrieved
during the API call. Calling API every 22 seconds
with large number of jobs bogs down the gui. Fixed
_getUpdate to include the root group and its'
associated jobs, this fn returned a different count
then the in mem ds. Fixed the recursive fn
__getNestedIds, had previously returned
only groups and not the associated jobs, so when
processUpdate compared the len() of what is in
self._items vs what was last retrieved ie
current == set(rpcObjects[1]) line 417, it was never
the same and continuously tried rebuilding self._items
for each interval.