Add a new property max_show_stale_days to control show expiration rule
The recently introduced PR#1151 ended up changing the default behaviour of the system, where shows would only
be marked as inactive manually. This PR keeps the functionally but adds a new property entry to set the number
of days a show needs to be stale (No new jobs) to be marked as inactive and sets the default to -1, which
deactivates the feature.