Revert booking balanced query for no-gpu
Using `between 0 and 0` seems to impact which query plan the database is choosing, causing up to 4x speed decrease in the worst case. The current setup chooses which query/arguments to use based on the number of idleGpus so there's some room for improvements in the future here.

This change only impacts the BALANCED scheduling mode as it is the most db intensive mode and seems to be the only one affected.

Signed-off-by: Diego Tavares <dtavares@imageworks.com>