Prevent booking frames on hosts with full MCP
Change the host hardwareState from REPAIR to UP, when the amount of free space in the /mcp directory is greater or equals to the minimum required and delete the comment associated with the host. Check if the host with REPAIR status has comments with subject=SUBJECT_COMMENT_FULL_MCP_DIR and user=CUEBOT_COMMENT_USER and delete the comments. If the comment exists, change the host hardwareState from REPAIR to UP.