Redesign free temp dir monitoring feature
The previous verison of this feature relied on a pre-defined limit for how much space needs to be available on the tmp directory. This posed a problem for non-homogeneous farms where different hosts can have great discrepancies on storage size.

This new feature relies on a configuring the minimun percentage os available temp storage.