Redesign free temp dir monitoring feature (#1394)
* Redesign free temp dir monitoring feature

The previous version of this feature relied on a pre-defined limit for how much space needs to be available on the tmp directory. This posed a problem for non-homogeneous farms where different hosts can have great discrepancies on storage size.

This new implementation relies on a pre-defined minimum percentage of available temporary storage.