Remove rqd restart feature (#1435)
The feature was never really functional. Its implementation relied on a
hardcoded path only valid for the init.d deployment of rqd, and even for
this use case it didn't work as expected.

For the three existing rqd deployment modes (docker, init.d and systemd)
restarting should be accomplished by calling stop and start
sequentially. Out of the three modes, only Docker is not handled by this
PR, as it requires external control of the rqd service on docker, and
service definitions are out of the scope of this repo.