Abstract logging into seperate classes (RQDLogger and LogReader) (#1429)
**Link the Issue(s) this Pull Request is related to.**
#1401 #1427 

**Summarize your change.**
This abstracts the reading of logs and writing log files. This makes it
simpler to do things like supporting other logging types and conversions
(eg. make sure things are ASCII) in a single class across all platforms.
This is a smaller version of #1401 with just the file logging
abstracted. It also implements the changes from #1427

Co-authored by @ramonfigueiredo and @lithorus

---------

Signed-off-by: Jimmy Christensen <Lithorus@gmail.com>
Co-authored-by: Ramon Figueiredo <rfigueiredo@imageworks.com>