[cuebot] Fix issue with auto-retrying killed frames
Auto-retrying frames killed automatically by the OOM-kill logic was not working as expected. rqd is currently not able to report exit_signal=9 when a frame is killed by the OOM logic. The current solution sets exitStatus to Dispatcher.EXIT_STATUS_MEMORY_FAILURE before killing the frame, this enables auto-retrying frames affected by the logic when they report with a frameCompleteReport.