Make output viewer feature generic
Make the previous logic implemented with itView in mind something that
can be reused by other studios using different viewers.

This can be accomplished by setting the viewer command pattern and
a regex to extract arguments from outputs. The config file has an
example of how to use it for itView.