[rqd] Set uid and gid when creating user
When a user has to be created with useradd, the user should have the same uid and gid of the user who created the job. This ensures directories and files created during the execution have the write permissions.

Not explicitly setting uid and gid causes problems for dockerized rqds trying to write to shared storage volumes mounted in bind mode.