[rqd] Set uid and gid when creating user for a job (#1480)
When a user has to be created with `useradd`, the user should have the
same `uid` and `gid` of the user who created the job. This ensures
directories and files created during the job execution have the same
permissions.

Not explicitly setting uid and gid causes problems for dockerized rqds
trying to write to shared storage volumes mounted in bind mode.

- This PR also contains a minor change to allow overriding the value of
`SP_OS` (Constant used to identify a host distro) using the config file.