[cuegui] Add Request Core Buttons to MenuActions (#1477)
This new feature adds a new menu action that opens a new `dialog` to
request more cores via email for a given job.

The `dialog` is an Email composed where the message is pre-populated
with relevant data about the job. Fields, like `to`, `cc` and `bcc`,
need to be filled manually. The content of the message is editable.



![image](https://github.com/user-attachments/assets/563157e6-c901-4a1d-8109-c43f804ec0be)

---------

Signed-off-by: Diego Tavares <dtavares@imageworks.com>
Co-authored-by: Roula O'Regan <roregan@imageworks.com>