[cuegui] Limit user actions on alien jobs (#1488)
**Rationale**
Users shouldn't be allowed by default to execute destructive actions on
jobs they don't own. For "superusers" there should be a simple way to
toggle this behavior off. This preference should be saved on the user
profile.

**Implementation details**
This feature checks user permissions before allowing a user to "eat,
kill, retry" jobs that they do not own.
The behavior can be overwritten at: "File -> Enable Job Interaction".
This gets saved to the config file and is disabled by
default.

---------

Co-authored-by: Roula O'Regan <roregan@imageworks.com>