[cuegui] Allow previewing checkpointed frames (#1491)
Some Digital Content Creation tools (DCCs) provide a preview of the
images being rendered using a local server. This feature allows
capturing a preview on a tmp directory for displaying using the
configured viewer.

For this to work, the host running the frame being previewed needs to
have the port being used by the DCC accessible. The hostname is used as
the URL and the port is captured from the frame logs, as DCCs with this
feature usually output the port in the first lines of the log file.

One last change on this PR is to make sure both Preview and View Output
options are only displayed when the viewer cmd pattern is configured on
`cuegui.yaml`.