Fix permission issues when becoming an user
Changes implemented by #1416 impacted the locking mechanism for handling permissions on rqd, causing multiple threads to compete for permission settings and access to passwd.

Besides fixing the bug, this PR also introduces a fix for a potential security issue that would allow frames to run as root if the frame user didn't exist and the process to create this user fails.