Update java on the base image for sonar-cloud
Ideally the `ci-opencue` image could be built with the correct version of the java jdk, but for now we're updating the version before running the pipeline to confirm it really works.

The only way to confirm this works as expected is merging to master as secret-keys are not shared with pipelines triggered by PRs.