[cuebot] Fix packaging pipeline (#1525)
Fix both the registry issues for images that rely on centos7 and the pip
package compatibility issues for python 3.6.

Some changes on PR #1416 didn't take into account that part of the
packaging pipeline relies on centos7, which doesn't support python3.7
without installing from source. On the next release, support to rhl7
might be dropped as we upgrade the pipeline to rely on rocky9.