[rqd] Fix issue on rqd when killing a frame that no longer exists (#1533)
When killing a frame on a rqd node that has restarted and lost track of
its running frames, the request should return a status to be handled
accordingly and trigger the lostProc logic. The current behavior leaves
the frame stuck at RUNNING until the job itself is killed or eaten.