[cuegui/pycue] Fix Local Booking widget (#1581)
This feature has been inactive on opencue since the beginning. Changes
to port from Ice to Grpc were not properly tested and this widget never
really worked.

Local Rendering if a feature that allows users to claim the ownership of
a host (their workstation) and assign a job to execute frames on that
host. This is very useful in situations where the farm is busy but user
workstations aren't.

To access the feature, right-click on a job/layer and select "Use local
cores..". On the opened widget, the user can select how much cores,
memory and gpu to allocate to execute cue jobs. When confirmed, cuebot
will start dispatching frames to that host.

---------

Signed-off-by: Diego Tavares <dtavares@imageworks.com>