[rqd/cuebot] Hard and Soft memory limits (#1589)
Currently, frames are created with minimal requirements, but limits are
not enforced. This PR implements soft and hard limits on RQD when
running on docker mode.

For more information about soft and hard limits
[read](https://docs.docker.com/engine/containers/resource_constraints/).

Limits are calculated using the minimum memory defined for the layer and
a multiplier that can be tuned at Dispatcher.java. Ideally, these values
should be extracted to opencue.properties, but they are being used in a
static context, and interpreted before the file is actually read.

---------

Signed-off-by: Diego Tavares <dtavares@imageworks.com>