(ci/cd) Drop support for cy2022 (#1603)
The ci-2022 image has been failing the pipelines for using an
unsupported version of GLIBC (see
https://github.com/AcademySoftwareFoundation/OpenCue/actions/runs/12016939050/job/33498210814)

According to the [Reference Platform](https://vfxplatform.com/) ci2022
uses **GLIBC_2.17**, but I'm not sure what on the pipeline requires a
higher version. Since this error happens on the `actions/checkout@v3`,
it looks like some version bump on GitHub's actions might be the root
cause.