[rqd] Add frame recovery logic for docker mode
Whenever rqd restarts it loses track of all the frames launched by it that haven't
finished. This change adds a new configurable option to backup frame states to a file,
that is used to recover the frame cache state and try to re-bind to the running frames.

This first version only works on docker mode