Fix redirect plugin table
The table on redirect plugin displays a count for 'pending frames'.
Pending takes into account frames that are not waiting for being
executed for having an active dependency. Before redirecting frames,
that is a check that will prevent redirects containing such frames.

This change makes it so the table displays 'waiting frames', which
takes into account only frames that are ready to be executed