Reformat all java files (#1627)
Formatting was previously being handled by code_style–ij.xml which is
not really compatible with many new IDEs. This PR changes formating to
GoogleStyle enforced by the spotless gradle plugin in conjunction with
jdtls, which is compatible with vscode and eclipse.

Execute `gradle spotlessApply` to execute the formatter using gradle
locally.