[cuegui] Fix slider performance issue on LayerDialog
Increasing the max_memory_limit on PR#1657 had a collateral effect
on LayerDialog causing int overflows and slowness on the Layer dialog.

The proposed solution is to stop using kb as the granularity of the
slider and start using MB. We lose percision, but I don't think people
would be that precious about it.