Check if uid is valid before launching frame
If the frame uid is outside of linux defined UID_MAX and UID_MIN
fallback to the uid of the daemon user. Read https://www.man7.org/linux/man-pages/man5/login.defs.5.html
for more information about uid limits. At this time, limits are
hardcoded on rqconstants as the linux default values, which sounds
reasonable as this default config limit is rarely modified.