[rqd] Change default RQD_BECOME_JOB_USER from True to False  (#1659)
**Summarize your change.**
Change default RQD_BECOME_JOB_USER from True to False

**Reason for change**
I think most studios don't expect the render client to change to the
user of the render submitter, but have a "render" user dedicated for it.
Also, it's not very obvious why root is required to run RQD and that the
behavior can be turned off.