[rqd] [cuegui] Add support for Loki for frame logs (#1577)
**Link the Issue(s) this Pull Request is related to.**
#1571 

**Summarize your change.**
This adds the ability to use loki as the backend for frame logs in rqd.
It also adds a new plugin/widget in cuegui to read the logs from the
loki server.

This enables logs files to not be bound by a single namespace and also
adds the potential to also store telemetry about the frame.

This is still on an experimental stage and any inputs are appreciated.
Screenshot of new widget :

![image](https://github.com/user-attachments/assets/aabd4e92-34c2-431f-a271-8211b7dbb579)

The loki-urllib3-client python module is optional and if it's not
detected, the cuegui widget will show and error message.

Not yet implemented :
- Search log
- Timestamp in log view

---------

Co-authored-by: Diego Tavares <dtavares@imageworks.com>