[cuegui] Change color of hosts with REBOOT_WHEN_IDLE status (#1666)
- Implement dark orange colour for REBOOT_WHEN_IDLE status
- Ensures high visibility and differentiation from other statuses

**Link the Issue(s) this Pull Request is related to.**

[#1665](https://github.com/AcademySoftwareFoundation/OpenCue/issues/1665)

**Summarize your change.**
Hosts with the LOCKED status are coloured blue, while hosts with the
DOWN or REPAIR statuses are coloured red.

This enhancement Implements a dark orange colour for hosts with the
REBOOT_WHEN_IDLE status to ensure high visibility and differentiation
from other statuses.

This change is based on the observation that when scanning the list of
hosts, we often don't look at the status. Therefore, hosts with the
REBOOT_WHEN_IDLE status might be lumped together with active hosts. The
colour is a highly noticeable indication that the host has a special
status.