[rqd] Rewrite rqnimby logic
The previous pynput based implementation was more complex than
needed due to the effort to keep backwards compatibility with the
previously implementation based on Select. This change removes
support to the Select based implementation in favor of a simplified
logic using pynput.

Given that the Select method was outdated and incompatible with modern
versions of linux, supporting it is no longer necessary.