Refactor Nimby logic to be retried on a schedule
For nimby to be able to import pynput, an user must be
signed in. This change keeps retrying to start rqnimby
on every 5 iterations of the main loop.

This also adds a configurable value for a path to a file
that can be sourced to fill up the DISPLAY environment
variable. The file should be written by an external script
(not part of this logic) that captures the signed in user
and writes its connected display (:0 for example) to the
file.