[rqd] Ensure logdir is owned by job user
Failing to do so causes problems if other frames from the same job
were executed on docker, which may lead to the log dir to be owned
by root.

This change attemps to change the logdir ownership prior to
initializing RqdLogger. Handling this internally on the initialization
logic would be preffered, but is not possible as the initalization
needs to run as the job user and the chown logic needs to happen on
permissionHigh mode.