Add build pipeline for packaging all OpenCue modules in CI workflow (#1785)
This is to fix the release pipeline.

It takes the same parts of the testing pipeline to first build the
python packages and use them for building the docker images.

I have tested the integration tests that seems to work, but I cannot
test the other steps, which requires AWS credentials.

This is needed for #1779 to work properly and generally to allow the
packaging to work again after the rework of the build setup for the
python libraries.