Prevent race condition trying to book the same host
There is a protection against processing multiple bookings on a single host at the same time on
HostDao that uses a database lock. This protection is intended for multiple instances of the
scheduler running at the same time. However, this logic was also being triggered by a single
instance, which indicated there was a race condition in place.

The race condition happens because hosts can belong to multiple groups at the same time.