[rust/rqd] Trigger a kill_all_frames action when a host becomes nimby_locked (#2057)
* To maintain feature parity with the previous Python/RQD, all frames
should be terminated whenever a
machine becomes nimby_locked. This ensures that the user has full access
to its resources.
To simplify the implementation of this cross-service action, some
services have adopted a
singleton-instance pattern. This pattern includes a public `instance`
method that provides access to
the live, lazily initialized instance of the service. The affected
services are frame_manager,
config, report_client, and machine_monitor.

* When trying to force_kill a frame, the first attempt uses
kill_session, and if it fails, it will try
force kill. There was a fail condition that could lead to an infinite
loop.

---------

Signed-off-by: Diego Tavares <dtavares@imageworks.com>